feat: add global storage to carti Global storage is a feature that is a slight requirement for supporting default packages. The storage implementation here uses the same constructs as everywhere else just adds a little extra searching in the cli layer vs core. This allows us to swap out in the future storage paradigms and change the logic in the cli as we see fit.
A usability global restriction was added to publish and bundle
while the code path exists to support publishing from global dir
and bundling from the global dir, this functionality is removed
from the cli.

Developing from a global dir is an easy way to footgun. So this
is rm'd. If we need it later for an explicit usecase we will add
it. You can work around it so it's not impossible.

This also introduces a build dir that can be opt'd out
in the future this may not be such an important feature.

However because much of cartesi is run in different context,
the ability to mount a unified path for the machine emulator is
invaluable. So carti_build/bundles is the location for all
bundles from an installed machine.json making it easy to
reference them in any context.